fetchAddProduct

+++ src/components/upLoad/upLoadWidget.js
@@ -14,20 +14,24 @@ export const UploadWidget = ({ ...props }) => {
   const cloudinaryRef = React.useRef();
   const widgetRef = React.useRef();
   const [userAvatarId, setUserAvatarId] = React.useState(null);
+  const isAdmin = props.admin;
+
 
   const cld = new Cloudinary({
     cloud: {
       cloudName: "dnd2lc6qw",
     },
   });
-  
+
   React.useEffect(() => {
-    if (authState.isLoading === 'loaded') {
-      setUserAvatarId(authState.data?.user.avatar);
-    }else{
-      dispatch(fetchAuthMe())
+    if (!props.admin) {
+      if (authState.isLoading !== 'loaded') {
+        dispatch(fetchAuthMe)
+      } else if (authState.isLoading === 'loaded') {
+        setUserAvatarId(authState.data?.user.avatar);
+      }
     }
-  }, [dispatch, authState.isLoading, authState.data?.user.avatar]);
+  }, [dispatch, authState.isLoading, authState.data?.user.avatar, props.admin]);
 
   const upload = useCallback(() => {
     cloudinaryRef.current = window.cloudinary;
@@ -43,43 +47,74 @@ export const UploadWidget = ({ ...props }) => {
     }, function (error, result) {
       try {
         const photoId = result.info.public_id;
-        // const avatarUrl = result.info.url;
-        if (photoId) {
-          const data = { email: authState?.data?.user?.email, avatar: photoId };
-          dispatch(fetchDeleteAvatar(userAvatarId))
-          setTimeout(() => {
-            dispatch(fetchUpdateAvatar({ ...data }));
-            dispatch(fetchAuthMe())
-          }, 400)
+        const productUrl = result.info.url;
+        if (!props.admin) {
+          if (photoId) {
+            const data = { email: authState?.data?.user?.email, avatar: photoId };
+            dispatch(fetchDeleteAvatar(userAvatarId))
+            setTimeout(() => {
+              dispatch(fetchUpdateAvatar({ ...data }));
+              dispatch(fetchAuthMe())
+            }, 400)
+          }
+        }
+        if (props.admin && productUrl) {
+          props.setUrl(productUrl)
         }
       } catch (e) {
         console.log(error);
       }
+
     });
     widgetRef.current.open()
-  },[dispatch, userAvatarId, authState?.data?.user?.email])
+  }, [dispatch, userAvatarId, authState?.data?.user?.email, props])
 
   const defaultAvatar = "https://res.cloudinary.com/dnd2lc6qw/image/upload/v1700142769/ha19yqibgjmmxvnfszbr.png";
+  const defaultProduct = "https://blockbusterbd.com/uploads/movies/posters/noimage.jpg";
   const userAvatar = cld.image(userAvatarId ? userAvatarId : defaultAvatar).format('auto').quality('auto');
+
+  const user = () => {
+    return (
+      <>
+        {
+          authState.data && authState.data.user.avatar ?
+            <AdvancedImage
+              className={s.accaunt__avatar}
+              onClick={() => upload()}
+              cldImg={userAvatar}
+            />
+            :
+            <img
+              className={s.accaunt__avatar}
+              src={defaultAvatar}
+              alt="avatar"
+              onClick={() => upload()}
+            />
+        }
+      </>
+    )
+  }
+
+  const admin = () => {
+    return (
+      <>
+        <div className={s.product} onClick={() => upload()}>
+          <img className={s.product__img} src={props.url ? props.url : defaultProduct} alt="productImage" />
+        </div>
+      </>
+    )
+  }
 
   return (
     <>
       {
-        authState.data && authState.data.user.avatar ?
-          <AdvancedImage
-            className={s.accaunt__avatar}
-            onClick={() => upload()}
-            cldImg={userAvatar}
-          />
+        !isAdmin ?
+          user()
           :
-          <img
-            className={s.accaunt__avatar}
-            src={defaultAvatar}
-            alt="avatar"
-            onClick={() => upload()}
-          />
+          admin()
       }
     </>
+
   )
 }
